add SegmentedButtons on AllKeyboard
