+++ hackathon/allKeyboard/components/Misc.tsx
@@ -1,0 +1,12 @@
+import { View, Text } from "react-native";
+import React from "react";
+
+interface Props {}
+
+export const Misc = (props: Props) => {
+  return (
+    <View>
+      <Text>Misc</Text>
+    </View>
+  );
+};
